shuffle partition changed to String (#717)
* shuffle partition changed to String

* comments implemented

* comments implemented

* comments implemented

* comments implemented

* comments implemented

* comments implemented

* comments implemented

* test cases added

* test cases added

+++ src/main/scala/com/databricks/labs/overwatch/pipeline/Initializer.scala
@@ -579,7 +579,7 @@ object Initializer extends SparkSessionWrapper {
     // set spark overrides in scoped spark session
     config.registerInitialSparkConf(spark.conf.getAll)
     config.setInitialWorkerCount(getNumberOfWorkerNodes)
-    config.setInitialShuffleParts(spark.conf.get("spark.sql.shuffle.partitions").toInt)
+    config.setInitialShuffleParts(spark.conf.get("spark.sql.shuffle.partitions"))
     if (debugFlag) {
       envInit("DEBUG")
       config.setDebugFlag(debugFlag)
